requirements.txt
streamlit==1.39.0
python-dotenv==1.0.1
google-generativeai==0.8.1
Pillow==10.4.0
requests==2.31.0